+++ utils/flexpart_ifs_utils/s3_utils.py
@@ -1,5 +1,4 @@
-#!/usr/bin/python3
-import boto3  # pip install boto3
+import boto3
 import os
 import argparse
 from datetime import datetime
@@ -54,3 +53,17 @@ if args.bucket is None:
     print("Skipping upload to S3 as S3_BUCKET_NAME environment variable is unset.")
 else:
     uploadDirectory(args.directory, args.bucket, args.endpoint_url)
+
+def download_file(obj, file):
+    bucket = os.getenv('FLEXPART_S3_BUCKETS__INPUT__NAME')
+    endpoint_url = os.getenv('FLEXPART_S3_BUCKETS__INPUT__ENDPOINT_URL')
+
+    s3 = boto3.client('s3', endpoint_url=endpoint_url, aws_access_key_id=os.getenv('S3_ACCESS_KEY'),
+            aws_secret_access_key=os.getenv('S3_SECRET_KEY'),)
+
+    try: 
+        with open(file, 'wb') as f:
+            print(f"Downloading {obj} to {f} from bucket: {bucket}")
+            s3.download_fileobj(bucket, obj, f)
+    except Exception as e:
+        raise(e)
